+++ src/backend/defs/go-objects/contact.go
@@ -627,31 +627,31 @@ func (lookup *ContactByContactPoints) UpdateLookups(contacts ...interface{}) fun
 				lkp := lookup.(*ContactByContactPoints)
 				lkpKey := lkp.UserID + lkp.Value + lkp.Type
 				// check if it's a new uri
-				if _, ok := oldLookups[lkpKey]; !ok || !update {
-					err := session.Query(`INSERT INTO contact_lookup (user_id, value, type, contact_id) VALUES (?,?,?,?)`,
+				err := session.Query(`INSERT INTO contact_lookup (user_id, value, type, contact_id) VALUES (?,?,?,?)`,
+					lkp.UserID,
+					lkp.Value,
+					lkp.Type,
+					lkp.ContactID,
+				).Exec()
+				if err != nil {
+					log.WithError(err).Warnf(`[CassandraBackend] UpdateLookups INSERT failed for user: %s, value: %s, type: %s`,
 						lkp.UserID,
 						lkp.Value,
-						lkp.Type,
-						lkp.ContactID,
-					).Exec()
-					if err != nil {
-						log.WithError(err).Warnf(`[CassandraBackend] UpdateLookups INSERT failed for user: %s, value: %s, type: %s`,
-							lkp.UserID,
-							lkp.Value,
-							lkp.Type)
-					}
+						lkp.Type)
+				}
+				if _, ok := oldLookups[lkpKey]; !ok || !update {
 					// do the job of updating participants and discussion related tables
-					err = associateURIWithContact(session, lkp)
+					err = updateURIWithContact(session, lkp, true)
 					if err != nil {
 						// TODO : handle multiple error returns to merge into one error or break immediately ?
 						log.WithError(err).Errorf("[Contact.UpdateLookups] associateURIWithContact failed with lookup %+v", lkp)
 					}
-					if update {
-						// remove keys in removedLookups,
-						// thus at end it will only hold remaining entries that are not in the new state
-						delete(removedLookups, lkpKey)
+				}
+				if update {
+					// remove keys in removedLookups,
+					// thus at end it will only hold remaining entries that are not in the new state
+					delete(removedLookups, lkpKey)
 
-					}
 				}
 			}
 			if len(removedLookups) > 0 {
@@ -671,7 +671,7 @@ func (lookup *ContactByContactPoints) UpdateLookups(contacts ...interface{}) fun
 							lookup.Type)
 					}
 					// do the job of updating participants and discussion related tables
-					err = dissociateURIFromContact(session, *lookup)
+					err = updateURIWithContact(session, lookup, false)
 					if err != nil {
 						// TODO : handle multiple error returns to merge into one error or break immediately ?
 					}
@@ -695,25 +695,10 @@ func (lookup *ContactByContactPoints) CleanupLookups(contacts ...interface{}) fu
 				return err
 			}
 			for _, lookup := range related {
-				ids := lookup["contact_ids"].([]gocql.UUID)
-				updated_ids := []string{}
-				for _, id := range ids {
-					if id.String() != contact.ContactId.String() { // keep only contact_ids that are not from the deleted contact
-						updated_ids = append(updated_ids, id.String())
-					}
-				}
-				if len(ids) == 0 || // we found an empty lookup: it should have been removed ! cleaning up
-					len(updated_ids) == 0 { // lookup had only one contact_ids and we just deleted it. cleaning up
+				contact_id := lookup["contact_id"].(gocql.UUID)
+				if contact.ContactId.String() == contact_id.String() {
+					log.Info("Deleting contact_lookup for user %s with value %s", lookup["user_id"], lookup["value"])
 					err := session.Query(`DELETE FROM contact_lookup WHERE user_id = ? AND value = ? AND type = ?`,
-						lookup["user_id"],
-						lookup["value"],
-						lookup["type"]).Exec()
-					if err != nil {
-						return err
-					}
-				} else if len(ids) != len(updated_ids) { // an id has been pop, need to update contact_lookup
-					err := session.Query(`UPDATE contact_lookup SET contact_ids = ? WHERE user_id = ? AND value = ? AND type = ?`,
-						updated_ids,
 						lookup["user_id"],
 						lookup["value"],
 						lookup["type"]).Exec()
@@ -728,9 +713,10 @@ func (lookup *ContactByContactPoints) CleanupLookups(contacts ...interface{}) fu
 	return nil
 }
 
-// associateURIWithContact is algorithm to update hashes in ParticipantHash table
+// updateURIWithContact is algorithm to update hashes in ParticipantHash table
 // to be able to link URI with contact when building discussions
-func associateURIWithContact(session *gocql.Session, lkp *ContactByContactPoints) error {
+// if associate == true, then lkp is associated with contact, else it is dissociated from contact
+func updateURIWithContact(session *gocql.Session, lkp *ContactByContactPoints, associate bool) error {
 	// lookup uris_hashes where lkp's uri is embedded
 	lkpUri := lkp.Type + ":" + lkp.Value
 	uriLookups, err := session.Query(`SELECT * FROM hash_lookup WHERE user_id = ? AND uri = ?`, lkp.UserID, lkpUri).Iter().SliceMap()
@@ -751,7 +737,7 @@ func associateURIWithContact(session *gocql.Session, lkp *ContactByContactPoints
 			if err.Error() == "not found" {
 				continue
 			} else {
-				log.WithError(err).Warnf("associateURIWithContact failed to query participant_hash for user_id = %s, kind = %s, key = %s", lkp.UserID, "uris", uriHash.Hash)
+				log.WithError(err).Warnf("updateURIWithContact failed to query participant_hash for user_id = %s, kind = %s, key = %s", lkp.UserID, "uris", uriHash.Hash)
 				continue
 			}
 		}
@@ -761,7 +747,7 @@ func associateURIWithContact(session *gocql.Session, lkp *ContactByContactPoints
 			if err.Error() == "not found" {
 				continue
 			} else {
-				log.WithError(err).Warnf("associateURIWithContact failed to query participant_hash for user_id = %s, kind = %s, key = %s", lkp.UserID, "participants", uriHash.Hash)
+				log.WithError(err).Warnf("updateURIWithContact failed to query participant_hash for user_id = %s, kind = %s, key = %s", lkp.UserID, "participants", uriHash.Hash)
 				continue
 			}
 		}
@@ -771,29 +757,35 @@ func associateURIWithContact(session *gocql.Session, lkp *ContactByContactPoints
 	}
 
 	// for each participant_hash :
-	//     - compute a new participant_hash with current contact_id
+	//     - compute a new participant_hash
 	//     - add new bijection uri_hash <-> new_participant_hash
+	//     - remove former bijection
 	for _, participantHash := range participantHashes {
-		newParticipantHash, err := ComputeNewParticipantHash(lkpUri, "contact:"+lkp.ContactID, participantHash)
+		var newParticipantHash ParticipantHash
+		var err error
+		newParticipantHash, err = ComputeNewParticipantHash(lkpUri, lkp.ContactID, participantHash, urisComponents[participantHash.Value], associate)
 		if err != nil {
-			log.WithError(err).Warnf("associateURIWithContact failed to compute new participant hash for uri <%s> and former participant hash : %+v", lkpUri, participantHash)
+			log.WithError(err).Warnf("updateURIWithContact failed to compute new participant hash for uri <%s> and former participant hash : %+v", lkpUri, participantHash)
+		} else {
+			err = StoreURIsParticipantsBijection(
+				session,
+				UUID(uuid.FromStringOrNil(lkp.UserID)),
+				participantHash.Value,
+				newParticipantHash.Key,
+				urisComponents[participantHash.Value],
+				newParticipantHash.Components,
+			)
+			if err != nil {
+				log.WithError(err).Warn("StoreURIsParticipantsBijection failed")
+			} else {
+				// remove former bijection
+				err = RemoveURIsParticipantsBijection(session, participantHash)
+				if err != nil {
+					log.WithError(err).Warn("RemoveURIsParticipantsBijection failed")
+				}
+			}
 		}
-		StoreURIsParticipantsBijection(
-			session,
-			UUID(uuid.FromStringOrNil(lkp.UserID)),
-			participantHash.Value,
-			newParticipantHash.Key,
-			urisComponents[participantHash.Value],
-			newParticipantHash.Components,
-		)
 	}
-
-	return nil
-}
-
-// dissociateURIFromContact is algorithm to update hashes in ParticipantHash table
-// to remove uri_hash <-> contact association
-func dissociateURIFromContact(session *gocql.Session, lkp ContactByContactPoints) error {
 
 	return nil
 }
@@ -809,6 +801,35 @@ func StoreURIsParticipantsBijection(session *gocql.Session, userId UUID, uriHash
 	// store participants_hash -> uris_hash
 	e2 := session.Query(`INSERT INTO participant_hash (user_id, kind, key, value, components, date_insert) VALUES (?,?,?,?,?,?)`,
 		userId, "participants", participantHash, uriHash, participantComponents, now).Exec()
+	switch {
+	case e1 != nil:
+		return e1
+	case e2 != nil:
+		return e2
+	}
+	return nil
+}
+
+// RemoveURIsParticipantsBijection delete uris_hash <-> participants_hash bijection
+// in participant_hash table
+func RemoveURIsParticipantsBijection(session *gocql.Session, participantHash ParticipantHash) error {
+	var first, second string
+	first = participantHash.Kind
+	if first == "participants" {
+		second = "uris"
+	} else {
+		second = "participants"
+	}
+	e1 := session.Query(`DELETE FROM participant_hash WHERE user_id = ? AND kind = ? AND key = ? AND value = ?`,
+		participantHash.UserId,
+		first,
+		participantHash.Key,
+		participantHash.Value).Exec()
+	e2 := session.Query(`DELETE FROM participant_hash WHERE user_id = ? AND kind = ? AND key = ? AND value = ?`,
+		participantHash.UserId,
+		second,
+		participantHash.Value,
+		participantHash.Key).Exec()
 	switch {
 	case e1 != nil:
 		return e1